PowerShell: Rename powershell_auth_options.py -> powershell_authentication_options.py
