+++ monkey/agent_plugins/exploiters/powershell/src/powershell_authentication_options.py
