+++ monkey/agent_plugins/exploiters/powershell/src/powershell_client.py
@@ -12,7 +12,7 @@ from urllib3 import connectionpool
 from common.credentials import Credentials, LMHash, NTHash, Password, Username, get_plaintext
 from infection_monkey.i_puppet import TargetHost
 
-from .powershell_auth_options import AuthenticationOptions
+from .powershell_authentication_options import AuthenticationOptions
 
 logger = logging.getLogger(__name__)
