+++ monkey/agent_plugins/exploiters/powershell/src/powershell_remote_access_client.py
@@ -20,7 +20,7 @@ from infection_monkey.exploit.tools import (
 from infection_monkey.exploit.tools.helpers import get_random_file_suffix
 from infection_monkey.i_puppet import TargetHost
 
-from .powershell_auth_options import get_auth_options
+from .powershell_authentication_options import get_auth_options
 from .powershell_client import PowerShellClient
 from .powershell_options import PowerShellOptions
